requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
python-hebrew-numbers==0.2.3
PyYAML==6.0
soupsieve==2.3.2.post1
